move yaml parser to separate modules
